reto2_microservices/orders_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic[email]==2.5.3
pika==1.3.2
python-dotenv==1.0.0
